Yeni dosya yapısı yapıldı. Kod güncellendi.
